Melhoria na apresentação das exceções
