+++ main.py
@@ -1,7 +1,6 @@
 import sys
 import utils
 from antlr4 import *
-from sty import fg, bg, ef, rs
 
 from trabalhoFinalMyListener import MyListener
 from arquivos_antlr.trabalhoFinalLexer import trabalhoFinalLexer
@@ -10,25 +9,24 @@ from arquivos_antlr.trabalhoFinalParser import trabalhoFinalParser
 if __name__ == '__main__':
     if not sys.argv[1].endswith('.py'):
         utils.lanca_excecao('A extensão do arquivo deve ser .py')
-        sys.exit()
+    else:
+        with open(sys.argv[1]) as file:
+            exp = ''.join(file.readlines())
+        
+        data = InputStream(exp)
+        
+        # Lexer
+        lexer = trabalhoFinalLexer(data)
+        tokens = CommonTokenStream(lexer)
+        
+        # Parser
+        parser = trabalhoFinalParser(tokens)
+        tree = parser.prog()
 
-    print('\n', '-'*15, 'Trabalho Final', '-'*15, '\n')
-
-    with open(sys.argv[1]) as file:
-        exp = ''.join(file.readlines())
+        # Listener
+        l = MyListener()
+        walker = ParseTreeWalker()
+        walker.walk(l, tree)
     
-    data = InputStream(exp)
-    
-    # Lexer
-    lexer = trabalhoFinalLexer(data)
-    tokens = CommonTokenStream(lexer)
-    
-    # Parser
-    parser = trabalhoFinalParser(tokens)
-    tree = parser.prog()
-
-    # Listener
-    l = MyListener()
-    walker = ParseTreeWalker()
-    walker.walk(l, tree)
+    utils.fecha_programa()
